Solve fds open on error in free ast functions
