+++ Makefile
@@ -53,6 +53,7 @@ SOURCES = \
 	$(SRC_DIR)/execute/cmd_path.c \
 	$(SRC_DIR)/execute/execute.c \
 	$(SRC_DIR)/execute/pipe_exec.c \
+	$(SRC_DIR)/execute/utils_exec.c \
 	$(SRC_DIR)/builtins/pwd.c \
 	$(SRC_DIR)/builtins/env.c \
 	$(SRC_DIR)/builtins/unset.c \
