+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: migugar2 <migugar2@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/20 00:02:21 by migugar2          #+#    #+#             */
-/*   Updated: 2025/09/26 17:16:28 by migugar2         ###   ########.fr       */
+/*   Updated: 2025/09/27 19:40:32 by migugar2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -198,21 +198,25 @@ int			ft_echo(t_cmd *cmd);
 int			ft_exit(t_shell *shell, char *argv[]);
 int			ft_export(t_shell *shell, char **argv);
 
-void			export_perror_identifier(t_shell *shell, char *argv);
+void		export_perror_identifier(t_shell *shell, char *argv);
 
 int			export_print_all(t_shell *shell);
 
 // execution
-void		wait_last_pid(t_shell *shell, pid_t pid);
-int			execute_ast(t_shell *shell, t_ast **node);
-int			execute_ast_pipe(t_shell *shell, t_ast **node);
-int			run_builtin_external(t_shell *shell, t_ast **ast, pid_t *pid);
-int			execute_ast_cmd(t_shell *shell, t_ast **cmd);
-int			is_builtin(char *cmd);
-void		debug_builtin(const char *cmd);
+void		seq_close(t_ast *ast, int in_fd, int out_fd);
+int			ft_dup2(int *oldfd, int newfd);
 
+void		wait_last_pid(t_shell *shell, pid_t pid);
+int			is_builtin(char *cmd);
+int			run_builtin_external(t_shell *shell, t_ast **ast, pid_t *pid);
 int			resolve_cmd_path(t_shell *shell, char **dst, const char *cmd);
 int			get_cmd_path(t_shell *s, char **dst, const char *cmd, char **paths);
+
+void		debug_builtin(const char *cmd);
+
+int			execute_ast_cmd(t_shell *shell, t_ast **cmd);
+int			execute_ast_pipe(t_shell *shell, t_ast **node);
+int			execute_ast(t_shell *shell, t_ast **node);
 
 // signals
