+++ source/execute/pipe_exec.c
@@ -6,42 +6,11 @@
 /*   By: migugar2 <migugar2@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/22 00:24:27 by gomandam          #+#    #+#             */
-/*   Updated: 2025/09/27 17:30:32 by migugar2         ###   ########.fr       */
+/*   Updated: 2025/09/27 20:49:42 by migugar2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-void	seq_close(t_ast *ast, int in_fd, int out_fd)
-{
-	if (ast == NULL || ast->type == AST_CMD || ast->type == AST_SUBSH)
-		return ;
-	if (ast->type == AST_PIPE)
-	{
-		if (ast->u_data.op.pipe_fd[0] != in_fd
-			&& ast->u_data.op.pipe_fd[0] != out_fd)
-			ft_close(&ast->u_data.op.pipe_fd[0]);
-		if (ast->u_data.op.pipe_fd[1] != in_fd
-			&& ast->u_data.op.pipe_fd[1] != out_fd)
-			ft_close(&ast->u_data.op.pipe_fd[1]);
-	}
-	seq_close(ast->u_data.op.left, in_fd, out_fd);
-	seq_close(ast->u_data.op.right, in_fd, out_fd);
-}
-
-int	ft_dup2(int *oldfd, int newfd)
-{
-	int	status;
-
-	if (*oldfd != newfd && *oldfd != STDIN_FILENO && *oldfd != STDOUT_FILENO)
-	{
-		status = dup2(*oldfd, newfd);
-		ft_close(oldfd);
-		if (status == -1)
-			return (-1);
-	}
-	return (0);
-}
 
 void	execute_seq_cmd(t_shell *shell, t_ast **cmd, int *in_fd, int *out_fd)
 {
@@ -143,7 +112,7 @@ int	execute_ast_pipe(t_shell *shell, t_ast **node)
 	count = (*node)->u_data.op.wait_count;
 	last = execute_seq_pipe(shell, node, &in_fd, &out_fd);
 	if (last == -1)
-		return (1); // TODO: still are open fds in wrong expansion or pipe error
+		return (1);
 	errno_cpy = errno;
 	wait_last_pid(shell, last);
 	while (count > 1)
